horus-taufiq-exam/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
mysql-connector-python==8.1.0
PyMySQL==1.1.0